Compatibility with Google Cloud Job Search/Discovery API
Unfortunately, there's a very significant problem with this, and that's
the unfortunate fact that this API works only with jobs that you submit
through the application. This means that the public jobs that I was
hoping to retrieve (which can be seen via google searches) are not as
easily available to retrieve as I hoped.

I'll do some more research into this; I sent in a request ticket that
Google may or may not respond to, but otherwise, it looks like I'll have
to deconstruct the huge framework that I just made, and focus on either
web scraping or other job website APIs, none of which I'm particularly
looking forward to working with. Ugh.