Rejection of machine learning in favor of nlp
Not much code added, but rather, most of the value came from roughly
2 hours of research as to how would be the best way to handle the
"keyword analysis" part of this application.

Originally, my thinking was to utilize tensorflow to search for
keywords. The problem with this strategy, though, is multipronged.

Firstly, training a machine learning model to classify words
individually is futile. Essentially, the model would be searching for
keywords that were explicitly defined to it as target keywords, since
looking at the word composition would be useless as to determining the
"keywordness" of a word. As such, any machine learning training would
require context; most likely, semantic recognition of keywords from
sentences or even the entire job description. That's something that's
still being researched on today.

Additionally, not only would machine learning training require a great
deal of resources (meaning I would either have to rent a computer or
force my poor laptop to work 24/7), but it would also require a large
dataset to train on. Obtaining the job descriptions might not be too
bad, but trying to sift through each description and marking key words?
That's basically slave labor; heck no.

So, instead, I will be using natural language processing, spearheaded by
nlp-compromise. Not only will it be used to correct for typos(?) and
converge different tenses and uses of the same word into the same
classification, but it might also be usable for faster detection for
keywords; namely, searching for adjectives, nouns, and verbs(?).

I will probably have to write a script to derive keywords, however,
so that's probably going to be my current focus. Adding additional job
APIs is a task that I can always do later, and who knows; maybe by the
time I finish, Google will have added the public attribute to jobs
stored in the Google cloud job search API, presenting me with a plethora
of additional resources to use.