Wrote word analysis algorithm, more plans
A lot of the time I spent today was on trying to decide the correct
route to go. I thought about using natural language processing in the
way that I described previously, but after a few trials, it didn't work
out. So, what I've done is created a basic word analysis algorithm that
provides as much inbuilt functionality as it can while also paving
the way for providing the user with as much information as possible.

I also came up with more ideas for additional features and
improving user experience. Looks like this project will take a while.
Ultimately, my goal is to host this on heroku so that others can use it
as well.